Extract shared chat mention picker hook and test util
Both ChatTray and CorpusChat duplicated ~90 lines of identical wiring for the
agent @mention picker: useAgentMentionTrigger + useUnifiedMentionSearch setup,
agentItems mapping, portaled AgentMentionPopover with rect-anchored positioning,
caret-restoring onSelect handler, and textarea onChange integration.

Pull all of that into useChatMentionPicker (frontend/src/hooks/) — each
consumer now calls a single hook and renders one popoverNode. Behaviour
preserved: portal anchoring to textareaRef, setTimeout-based caret restore,
conditional rendering when picker is open + has agents.

Also extract the identical buildMentionSearchMocks Apollo MockedResponse
builder from ChatTray.ct.tsx and CorpusChat.ct.tsx into
frontend/tests/utils/mentionSearchMocks.ts so the two surfaces stay in
lockstep when useUnifiedMentionSearch adds resource categories.

All 62 ChatTray + CorpusChat CT tests still pass.
